requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
Django==2.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytest==4.4.0
pytest-django==3.4.8
pytz==2018.9
six==1.12.0
